requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
requests==2.31.0
pillow==10.1.0
beautifulsoup4==4.12.2
asyncpg==0.29.0
motor==3.3.2
pymongo==4.6.0
